Small Update: Added new Shield Timer
Visual representation on the bottom of the screen of the shield on how
long it lasts, as well as how long it takes to be ready, as well as a
color swap when it is ready. Currently, 200 hits total will give the
player a full shield after the Cooldown for using it is up (set to 5
seconds).
Need to remove old shield timer from top of the screen, as well as add
UI elements to replace the old text based ones.